refactor: set domains
