+++ README.md
@@ -15,7 +15,7 @@ function applyTimeMask(e) {
   // 13    OUTPUT 00h13
   // 130   OUTPUT 01h30
   // 13000 OUTPUT 13h00
-  e.currentTarget.value = maskWizard.interval(e.currentTarget.value);
+  e.currentTarget.value = maskWizard.time.interval(e.currentTarget.value);
 }
 
 <input placeholder="Example" onChange={applyTimeMask} />;
